blueprints/multi-agent-compliance-analysis/backend/agents/lawyer_agent/pyproject.toml
[project]
name = "lawyer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-lambda-powertools>=3.19.0",
    "bedrock-agentcore>=0.1.2",
    "boto3>=1.28.0",
    "botocore>=1.31.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-aws>=0.2.26",
    "langchain-core>=0.3.66",
    "opensearch-py>=3.0.0",
    "pydantic>=2.11.7",
    "strands-agents>=1.5",
    "strands-agents-tools>=0.2.4",
    "uvicorn[standard]>=0.35.0",
]
